+++ ChangeLog.markdown
@@ -16,7 +16,7 @@ Made the following enhancements to improve R<sup>7</sup>RS support:
   in addition: "Strings and symbols now allow mnemonic and numeric escape sequences, and the list of named characters has been extended."
 - WIP: log second argument (base)
 - Added `boolean=?`, `symbol=?`, `read-line`, `eof-object`
-- Added `letrec*`, `when`, and `unless` syntax, per r7rs.
+- Added `include`, `letrec*`, `when`, and `unless` syntax, per r7rs.
 
 v3.13
 --------